Remove auto-inferring releases from git tags.
This is going to break support for non-Rust projects, but it's neccesary
for now. Basically, 1.12.1 screws things up: with the "let's look at the
tags" version, 1.13.0 is compared against 1.12.1, but that's wrong: it
needs to be compared against 1.12, due to the way branches work. If we
do this, then all the 1.12.1 commits get assigned to 1.13.0.

The regression is unfortunate but getting it right is more important for
now.