Test static file existence with .is_file() instead of .exists()
`.exists()` passes for directories, which then crash the app at
`let mut f = File::open(&path[1..]).unwrap()`. `.is_file()` is false both when
the file doesn't exist and if it's a path to a directory.

An example of a local URL that could crash the app before this change:
http://localhost:1337/public/fonts

I hope that production is running behind a supervisor that would restart it,
but I didn't test this on the live site to find out :P