Change `authors::load_or_create` to be less racy
This adds a unique constraint on authors (which appears to have been
implicit before, but not verified), and uses upsert to be a bit less
racy. I would have put this as an instance method on `NewAuthor`, but
the code seems to be structured to avoid that as a matter of taste so I
didn't do that restructuring.